requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
langchain_groq
langchain_chroma
langchain_huggingface
langchain-text-splitters
langchain
streamlit
sentence-transformers
pypdf
faiss-cpu
duckduckgo-search
ddgs
pymupdf
arxiv
wikipedia